Add ace support
Can you try to fix / change how the camera is handled?  RN it alway glitches through the player cause the eyes are not a consistent reference point for a camera. You can't even adjust the camera correctly cause the adjustment depends on the DIR and that changes if you look in North East South West. Just try it out, the cam is a bit broken IDK how to fix that.